pyproject.toml
[project]
name = "video-subtitler"
version = "1.0.0"
description = "Wrapper of Whisper OpenAI to generate .srt files for videos"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "openai-whisper>=20240930",
    "pydub>=0.25.1",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.51.3",
]
